edit for ribbon clients

+++ Core_Role/src/main/java/de/hska/iwi/vslab/Core_Role/Interfaces/RoleRepository.java
@@ -12,7 +12,7 @@ public interface RoleRepository extends CrudRepository<Role, Long> {
 
     List<Role> findAll();
 
-    long deleteById(int id);
+    int deleteById(int id);
 
-    long deleteByType(String type);
+    int deleteByType(String type);
 }
